Add files via upload
This code includes the main user-interface, or maybe user-input of the program. However, this code needs an attachment module; which is PlotPly. 

If you want to download this code, the first essential step is go to Command Prompt or CMD. Then, type in "pip install PlotPly". When done, the code should execute smoothly. 

Attachments - PlotPly module